Use singleton scope instead of request scope
Request scope causes problems with unit testing because the scope
`request` is not valid outside of a web context. Further, Fedora
integration testing doesn't work properly with this scope yet.

Singleton scope is fine because the SWORDServiceResource is ment to be
stateless. Multi threading is the default state of operation for the
Spring dispatcher anyway, so there is no need for request scoping.